+++ flukit/config/mutations_h1n1.tsv
@@ -1,9 +1,24 @@
 gene	pos
 PA	22
-PA	23
-PA	35
 PA	37
-PA	118
 PA	198
 NA	274
 MP	30
+PA	197
+PA	36
+PA	33
+NA	109
+NA	116
+NA	118
+NA	135
+NA	151
+NA	198
+NA	222
+NA	246
+NA	292
+NA	294
+NA	426
+NA	435
+NA	457
+NA	146
+NA	312
