+++ flukit/config/mutations_h3n2.tsv
@@ -1,9 +1,28 @@
 gene	pos
 PA	22
-PA	23
 PA	35
 PA	37
 PA	118
 PA	198
-NA	273
+NA	274
 MP	30
+PA	27
+PA	33
+PA	36
+PA	197
+NA	118
+NA	135
+NA	150
+NA	221
+NA	223
+NA	273
+NA	291
+NA	328
+NA	330
+NA	333
+NA	370
+NA	390
+NA	147
+NA	244
+NA	248
+NA	293
