--- mutations_vic.tsv
@@ -1,47 +1,0 @@
-gene	pos
-NA	272
-PA	22
-PA	35
-PA	37
-PA	198
-PA	33
-PA	119
-NA	100
-NA	103
-NA	104
-NA	107
-NA	116
-NA	133
-NA	137
-NA	138
-NA	139
-NA	141
-NA	144
-NA	145
-NA	149
-NA	150
-NA	151
-NA	168
-NA	196
-NA	199
-NA	220
-NA	244
-NA	245
-NA	246
-NA	291
-NA	359
-NA	360
-NA	373
-NA	394
-NA	395
-NA	406
-NA	431
-NA	438
-NA	463
-NA	105
-NA	164
-NA	185
-NA	261
-NA	272
-NA	293
-NA	143
